Add files via upload
Fuel Economy Data Analysis
This project is a data analysis of fuel economy data provided by the Environmental Protection Agency (EPA) for the years 2008 and 2018. The dataset contains information on the fuel economy, greenhouse gas emissions, and other relevant features of various car models.

The goal of this project is to explore and compare the trends in fuel economy and emissions between the two years, and to identify the characteristics of cars with the best fuel economy. The analysis includes data cleaning and wrangling, exploratory data analysis, and visualization using Python libraries such as Pandas, NumPy, and Matplotlib.

The project also includes a Jupyter notebook with the code used for the analysis, as well as a detailed report of the findings and conclusions. This project can be useful for anyone interested in the fuel economy and environmental impact of cars, and can provide insights for policymakers and car manufacturers.